socket/requirements.txt
dnspython==2.7.0
pymongo==4.10.1
websocket-client==1.8.0
